Fix "os-prober: call dmraid -r *once*"
The stdout of the command type is redirected >/dev/null that causes the
variable DMRAID to be always empty.

Therefore, the following if statement is always false and raid devices
are never listed to the file dmraid-map.

Furthermore, the command dmraid exists with failure if raid devices and
print the message "no raid disks" to stdout.

This commit fixes 9d89a52514897014df8c9d7c62a2985708bec618.